Nova feature(inacabada): método para buscar por idade

+++ digital/src/main/java/hospital/digital/entity/paciente/Paciente.java
@@ -54,7 +54,7 @@ public class Paciente {
     public Paciente() {
     }
 
-    public int calcularIdade(){
-        return Period.between(data_nasc, LocalDate.now()).getYears();
+    public byte calcularIdade(){
+        return (byte) Period.between(data_nasc, LocalDate.now()).getYears();
     }
 }
